requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
python-dotenv==1.0.0
Werkzeug==2.3.7
Markdown==3.4.4
Flask-WTF==1.1.1
WTForms
requests
Flask-Migrate
Flask-Babel
Flask-Moment
openai
gunicorn
